letsur-serving-app-fastapi/pyproject.toml
[project]
name = "letsur-serving-app-fastapi"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "celery[s3,sqs]==5.4.0",
    "email-validator==2.1.1",
    "fastapi[standard]==0.115.12 ; python_full_version >= '3.9' and python_full_version < '4.0'",
    "fsspec==2024.6.0",
    "pillow==10.3.0",
    "pip>=25.1.1",
    "pydantic-settings==2.4",
    "pylibmagic==0.5.0",
    "python-magic==0.4.27",
    "pytz==2024.2",
    "redis==5.0.8",
    "s3fs==2024.6.0",
    "uvicorn[standard]==0.29.0 ; python_full_version >= '3.9' and python_full_version < '4.0'",
    "langfuse==2.54.1",
]



[build-system]
requires = ["setuptools>=61", "setuptools-scm"]
build-backend = "setuptools.build_meta"


[tool.setuptools_scm]
version_file = "src/_version.py"

[dependency-groups]
dev = [
    "setuptools-scm>=8.3.1",
    "tomlkit>=0.13.3",
]
